Max priority queue fix

+++ Source/EasyNetQ/IAdvancedBus.cs
@@ -259,7 +259,7 @@ namespace EasyNetQ
             bool autoDelete = false,
             int? perQueueMessageTtl  = null, 
             int? expires = null,
-            byte? maxPriority = null,
+            int? maxPriority = null,
             string deadLetterExchange = null, 
             string deadLetterRoutingKey = null);
 
@@ -285,7 +285,7 @@ namespace EasyNetQ
             bool autoDelete = false,
             int? perQueueMessageTtl  = null,
             int? expires = null,
-            byte? maxPriority = null,
+            int? maxPriority = null,
             string deadLetterExchange = null, 
             string deadLetterRoutingKey = null);
